Match empty braces as increasing indent
This is required for TextMate’s auto-indent to behave correct, as when we press return after ‘if’ the indent is increased by one, i.e.:

    if(condition)
        ‸

If the user type a single ‘{’ with auto-pairing of braces we end up with:

    if(condition)
        {‸}

Here we should match that construct as “increasing indent” so that it gets de-indented (cancels the indent from the ‘if’ above).

Fixes #16.